Move inline Git blame into the GitCoordinator
The blame annotation engine — eligibility, the off-thread per-file
fetch, and the blame→BlameInfo mapping (author/date label, full-commit
tooltip, age-heatmap tint) — moves out of MainController into
GitCoordinator, where applyState drives it directly. That removes the
WindowOps.refreshBlame Host hop added when the engine was extracted:
the coordinator now owns blame end to end and needs nothing new from
the window (only the shared CoordinatorHost).

The click→commit-diff actions (blameShowCommit/onGutterBlameClick)
stay in MainController since they open a diff tab. No behavior change;
full suite green (1072).
